WIP: moving classes and organizing
